+++ site/jaws_site/task_logger.py
@@ -1,21 +1,14 @@
-import json
 import logging
 from datetime import datetime
 
-import pika
+from jaws_common.exceptions import JawsDbUnavailableError
 from sqlalchemy.exc import IntegrityError, OperationalError, SQLAlchemyError
 
 from jaws_site import models
 
-DATETIME_FMT = "%Y-%m-%d %H:%M:%S"
-
-
-class TaskLoggerDbError(Exception):
-    pass
-
 
 class TaskLogger:
-    def __init__(self, config, session, logger=None) -> None:
+    def __init__(self, config, session, **kwargs) -> None:
         """
         The task logger receives RabbitMQ messages and saves them in the RDb.
         :param config: Configuration parameters
@@ -25,24 +18,21 @@ class TaskLogger:
         """
         self.config = config
         self.session = session
-        if logger is None:
+        self.logger = kwargs.get("logger", None)
+        if self.logger is None:
             self.logger = logging.getLogger(__package__)
-        else:
-            self.logger = logger
 
-    def save(self, message: str) -> bool:
-        params = json.loads(message)
+    def save(self, params: dict) -> bool:
         timestamp = params.get("timestamp")
         params["timestamp"] = datetime.strptime(timestamp, "%Y-%m-%d %H:%M:%S")
         status = params.get("status")
+        if status not in ["queued", "running", "done"]:
+            raise ValueError(f"Invalid status: {status}")
 
         if status == "queued":
-            return self._insert(**params)
-        elif status in ("running", "done"):
-            return self._update(**params)
+            self._insert(**params)
         else:
-            self.logger.error(f"Invalid log status: {status}")
-            return True
+            self._update(**params)
 
     def _insert(self, **kwargs) -> bool:
         """
@@ -63,16 +53,17 @@ class TaskLogger:
         except OperationalError as error:
             # this is the only case in which we would not want to ack the message
             self.logger.error(f"Unable to connect to db: {error}")
-            return False
+            raise JawsDbUnavailableError(str(error))
         except IntegrityError as error:
             self.session.rollback()
             self.logger.error(f"Invalid task-log message, {kwargs}: {error}")
+            raise
         except SQLAlchemyError as error:
             self.session.rollback()
             self.logger.exception(
                 f"Failed to insert task log for Task {cromwell_run_id} {task_dir}: {error}"
             )
-        return True
+            raise
 
     @staticmethod
     def delta_minutes(start, end) -> int:
@@ -110,16 +101,15 @@ class TaskLogger:
         except OperationalError as error:
             # this is the only case in which we would not want to ack the message
             self.logger.error(f"Unable to connect to db: {error}")
-            return False
+            raise JawsDbUnavailableError(str(error))
         except Exception as error:
             err_msg = f"Unexpected error; unable to select task log {cromwell_run_id} {task_dir}: {error}"
             self.logger.error(err_msg)
-            return True
+            raise
 
         if row is None:
             # this can only happen if the queued message was lost
             self.logger.error(f"Task {cromwell_run_id} {task_dir} not found!")
-            return True
 
         try:
             if status == "running":
@@ -135,51 +125,10 @@ class TaskLogger:
             # this is the only case in which we would not want to ack the message
             self.session.rollback()
             self.logger.error(f"Unable to connect to db: {error}")
-            return False
+            raise JawsDbUnavailableError(str(error))
         except SQLAlchemyError as error:
             self.session.rollback()
             self.logger.exception(
                 f"Unable to update Tasks {cromwell_run_id} {task_dir}: {error}"
             )
-        return True
-
-    def receive_messages(self):
-        """
-        Receive and consume task-log messages indefinately.
-        """
-        host = self.config.get("RMQ", "host")
-        port = self.config.get("RMQ", "port")
-        vhost = self.config.get("RMQ", "vhost")
-        user = self.config.get("RMQ", "user")
-        password = self.config.get("RMQ", "password")
-        ttl = int(
-            self.config.get("RMQ", "ttl", 3600000)
-        )  # must match value in sender's self.config
-        site_id = self.config.get("SITE", "id")
-        queue = f"{site_id}_tasks"
-
-        def callback(ch, method, properties, body):
-            message = body.decode()
-            self.logger.debug(str(message))
-            if self.save(message) is True:
-                ch.basic_ack(delivery_tag=method.delivery_tag)
-
-        credentials = pika.PlainCredentials(user, password)
-        connection = pika.BlockingConnection(
-            pika.ConnectionParameters(host, port, vhost, credentials)
-        )
-        channel = connection.channel()
-
-        try:
-            channel.queue_declare(
-                queue=queue, durable=True, arguments={"x-message-ttl": ttl}
-            )
-        except Exception as error:
-            self.logger.warning(f"Unable to declare RMQ queue: {error}")
-            # channel.queue_delete(queue=queue)
-            # connection.close()
             raise
-
-        channel.basic_consume(queue=queue, on_message_callback=callback, auto_ack=False)
-        self.logger.debug("Waiting for task-log messages")
-        channel.start_consuming()