+++ site/jaws_site/transfers.py
@@ -9,7 +9,6 @@ import os
 from datetime import datetime
 
 import boto3
-import botocore
 from parallel_sync import rsync
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -24,9 +23,9 @@ MAX_ERROR_STRING_LEN = 1024
 
 def mkdir(path, mode=None):
     if mode is None:
-        mode = int(config.conf.get("SITE", "folder_permissions"), base=8)
+        mode = int(config.conf.get("SITE", "folder_permissions", "775"), base=8)
     if os.path.isdir(path):
-        return
+        os.chmod(path, mode)
     else:
         (head, tail) = os.path.split(path)
         mkdir(head, mode)
@@ -175,50 +174,44 @@ class Transfer:
         """
         logger.debug(f"Begin transfer {self.data.id}")
         self.update_status("transferring")
+        result = None
+        reason = None
         try:
             if self.data.src_base_dir.startswith("s3://"):
                 self.s3_download()
             elif self.data.dest_base_dir.startswith("s3://"):
                 self.s3_upload()
             else:
-                self.local_rsync()
+                self.local_copy()
         except Exception as error:
-            self.update_status("failed", str(error))
+            result = "failed"
+            reason = str(error)
         else:
-            self.update_status("succeeded")
+            result = "succeeded"
+
+        # session may be stale, so close it to get a new connection
+        self.session.remove()
+        self.update_status(result, reason)
 
     def aws_s3_resource(self):
         aws_access_key_id = config.conf.get("AWS", "aws_access_key_id")
         aws_secret_access_key = config.conf.get("AWS", "aws_secret_access_key")
         aws_region_name = config.conf.get("AWS", "aws_region_name")
-        try:
-            aws_session = boto3.Session(
-                aws_access_key_id=aws_access_key_id,
-                aws_secret_access_key=aws_secret_access_key,
-                region_name=aws_region_name,
-            )
-        except Exception as error:
-            raise (f"Error getting aws session: {error}")
-        try:
-            s3_resource = aws_session.resource("s3")
-        except Exception as error:
-            raise (f"Error getting s3 sources: {error}")
-        else:
-            return s3_resource
+        aws_session = boto3.Session(
+            aws_access_key_id=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+            region_name=aws_region_name,
+        )
+        return aws_session.resource("s3")
 
     def aws_s3_client(self):
-        try:
-            aws_access_key_id = config.conf.get("AWS", "aws_access_key_id")
-            aws_secret_access_key = config.conf.get("AWS", "aws_secret_access_key")
-            client = boto3.client(
-                "s3",
-                aws_access_key_id=aws_access_key_id,
-                aws_secret_access_key=aws_secret_access_key,
-            )
-        except Exception as error:
-            raise (f"Failed to get AWS client: {error}")
-        else:
-            return client
+        aws_access_key_id = config.conf.get("AWS", "aws_access_key_id")
+        aws_secret_access_key = config.conf.get("AWS", "aws_secret_access_key")
+        return boto3.client(
+            "s3",
+            aws_access_key_id=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+        )
 
     @staticmethod
     def s3_parse_path(full_path):
@@ -226,7 +219,6 @@ class Transfer:
         folders = full_path.split("/")
         s3_bucket = folders.pop(0)
         path = "/".join(folders)
-        # print(f"S3 BUCKET={s3_bucket}; PATH={path}")
         return s3_bucket, path
 
     def s3_file_size(self, bucket, file_key, aws_s3_client=None):
@@ -241,20 +233,8 @@ class Transfer:
         :return: size
         :rtype: int
         """
-        if aws_s3_client is None:
-            aws_s3_client = self.aws_s3_client()
-        try:
-            result = aws_s3_client.list_objects_v2(Bucket=bucket, Prefix=file_key)
-        except botocore.exceptions.ClientError as error:
-            logger.error(f"Error getting S3 obj stats for {file_key}: {error}")
-            raise
-        except botocore.exceptions.ParamValidationError as error:
-            raise ValueError(
-                "The parameters you provided are incorrect: {}".format(error)
-            )
-        except Exception as error:
-            logger.error(f"Error getting S3 obj stats for {file_key}: {error}")
-            raise
+        aws_s3_client = self.aws_s3_client()
+        result = aws_s3_client.list_objects_v2(Bucket=bucket, Prefix=file_key)
         size = None
         if "Contents" in result:
             contents = result["Contents"]
@@ -276,74 +256,29 @@ class Transfer:
         manifest = self.manifest()
         num_files = len(manifest)
         logger.debug(f"Transfer {self.data.id} begin s3 upload of {num_files} files")
-        try:
-            aws_client = self.aws_s3_client()
-        except Exception as error:
-            # do not fail, retry later
-            logger.error(f"{error}")
-            return
-        try:
-            aws_s3_resource = self.aws_s3_resource()
-        except Exception as error:
-            # do not fail, retry later
-            logger.error(f"{error}")
-            return
-        try:
-            s3_bucket, dest_base_dir = self.s3_parse_path(self.data.dest_base_dir)
-        except Exception as error:
-            msg = f"Error parsing s3 uri, {self.data.dest_base_dir}: {error}"
-            logger.error(msg)
-            self.update_status("upload failed", msg)
-            raise
-        try:
-            bucket_obj = aws_s3_resource.Bucket(s3_bucket)
-        except Exception as error:
-            msg = f"Error accessing bucket, {s3_bucket}: {error}"
-            logger.error(msg)
-            self.update_status("upload failed", msg)
-            raise
+        aws_client = self.aws_s3_client()
+        aws_s3_resource = self.aws_s3_resource()
+        s3_bucket, dest_base_dir = self.s3_parse_path(self.data.dest_base_dir)
+        bucket_obj = aws_s3_resource.Bucket(s3_bucket)
         for rel_path in manifest:
             src_path = os.path.normpath(os.path.join(self.data.src_base_dir, rel_path))
             src_file_size = os.path.getsize(src_path)
             dest_path = os.path.normpath(os.path.join(dest_base_dir, rel_path))
-            try:
-                dest_file_size = self.s3_file_size(s3_bucket, dest_path, aws_client)
-            except Exception as error:
-                self.update_status("upload failed", f"{error}")
-                raise
+            dest_file_size = self.s3_file_size(s3_bucket, dest_path, aws_client)
             if dest_file_size is not None and src_file_size == dest_file_size:
-                logger.debug(f"S3 upload: Skipping cached file {dest_path}")
+                logger.debug(f"S3 upload: Skipping existing file {dest_path}")
             else:
                 logger.debug(f"S3 upload to {s3_bucket}: {src_path} -> {dest_path}")
-                try:
-                    with open(src_path, "rb") as fh:
-                        bucket_obj.upload_fileobj(fh, dest_path)
-                except Exception as error:
-                    msg = f"Failed to upload to S3, file {src_path} -> {dest_path}: {error}"
-                    logger.error(msg)
-                    self.update_status("upload failed", msg)
-                    raise IOError(error)
+                with open(src_path, "rb") as fh:
+                    bucket_obj.upload_fileobj(fh, dest_path)
 
     def s3_download(self):
-        return (
-            self.s3_download_files()
-            if len(self.manifest())
-            else self.s3_download_folder()
-        )
-
-    def s3_download_files(self):
         manifest = self.manifest()
         num_files = len(manifest)
         logger.debug(f"Transfer {self.data.id} begin s3 download of {num_files} files")
         aws_s3_resource = self.aws_s3_resource()
         s3_bucket, src_base_dir = self.s3_parse_path(self.data.src_base_dir)
-        try:
-            bucket_obj = aws_s3_resource.Bucket(s3_bucket)
-        except Exception as error:
-            msg = f"Error accessing bucket, {s3_bucket}: {error}"
-            logger.error(msg)
-            self.update_status("download failed", msg)
-            raise
+        bucket_obj = aws_s3_resource.Bucket(s3_bucket)
         for rel_path in manifest:
             src_path = os.path.normpath(os.path.join(src_base_dir, rel_path))
             dest_path = os.path.normpath(
@@ -353,60 +288,11 @@ class Transfer:
                 f"S3 download from {s3_bucket}: {rel_path} as {src_path} -> {dest_path}"
             )
             dest_folder = os.path.dirname(dest_path)
-            try:
-                mkdir(dest_folder)
-            except IOError as error:
-                msg = f"Unable to make download dir, {dest_folder}: {error}"
-                logger.error(msg)
-                self.update_status("download failed", msg)
-            try:
-                with open(dest_path, "wb") as fh:
-                    bucket_obj.download_fileobj(src_path, fh)
-            except Exception as error:
-                msg = f"Failed to download s3 file, {src_path}: {error}"
-                logger.error(msg)
-                self.update_status("download failed", msg)
-                raise IOError(error)
+            mkdir(dest_folder)
+            with open(dest_path, "wb") as fh:
+                bucket_obj.download_fileobj(src_path, fh)
 
-    def s3_download_folder(self):
-        aws_s3_client = self.aws_s3_client()
-        s3_bucket, src_base_dir = self.s3_parse_path(self.data.src_base_dir)
-        paginator = aws_s3_client.get_paginator("list_objects_v2")
-        for page in paginator.paginate(Bucket=s3_bucket, Prefix=src_base_dir):
-            for obj in page["Contents"]:
-                rel_path = obj["Key"]
-                size = obj["Size"]
-                dest_rel_path = rel_path.removeprefix(src_base_dir)
-                dest_path = os.path.normpath(
-                    os.path.join(self.data.dest_base_dir, f"./{dest_rel_path}")
-                )
-                logger.debug(f"S3 download {rel_path} -> {dest_path}")
-                if rel_path.endswith("/") and size == 0:
-                    try:
-                        mkdir(dest_path)
-                    except Exception as error:
-                        msg = f"Unable to make download dir, {dest_path}: {error}"
-                        logger.error(msg)
-                        self.update_status("download failed", msg)
-                        raise IOError(msg)
-                else:
-                    try:
-                        basedir = os.path.dirname(dest_path)
-                        mkdir(basedir)
-                    except Exception as error:
-                        msg = f"Unable to make download dir, {basedir}: {error}"
-                        logger.error(msg)
-                        self.update_status("download failed", msg)
-                        raise IOError(msg)
-                    try:
-                        aws_s3_client.download_file(s3_bucket, rel_path, dest_path)
-                    except Exception as error:
-                        msg = f"S3 download error, {rel_path}: {error}"
-                        logger.error(msg)
-                        self.update_status("download failed", msg)
-                        raise IOError(msg)
-
-    def local_rsync(self) -> None:
+    def local_copy(self) -> None:
         """
         Copy files and folders (recursively).
         """
@@ -416,11 +302,7 @@ class Transfer:
         if not os.path.isdir(src):
             raise FileNotFoundError(f"Source directory not found: {src}")
         dest = f"{self.data.dest_base_dir}/"
-        try:
-            mkdir(dest)
-        except IOError as error:
-            logger.error(f"Transfer {self.data.id} failed: {error}")
-            raise IOError(f"Transfer {self.data.id} failed: {error}")
+        mkdir(dest)
         rel_paths = abs_to_rel_paths(src, get_abs_files(src, manifest))
 
         num_files = len(rel_paths)
@@ -428,7 +310,7 @@ class Transfer:
         logger.debug(
             f"Transfer {self.data.id}: Copy {num_files} files using {parallelism} threads"
         )
-        parallel_rsync_files_only(rel_paths, src, dest, parallelism=parallelism)
+        parallel_copy_files_only(rel_paths, src, dest, parallelism=parallelism)
 
         logger.debug(f"Transfer {self.data.id}: Chmod files")
         file_mode = int(config.conf.get("SITE", "file_permissions"), base=8)
@@ -533,9 +415,9 @@ def calculate_parallelism(num_files):
     return max(parallelism, min_threads)
 
 
-def parallel_rsync_files_only(manifest: list, src: str, dest: str, **kwargs):
+def parallel_copy_files_only(manifest: list, src: str, dest: str, **kwargs):
     """
-    Given list of files, copy them in parallel using rsync.  Copies regular files only, skips others.
+    Given list of files, copy them in parallel using parallel_sync.  Copies regular files only, skips others.
     :param manifest: list of file relative paths
     :ptype manifest: list
     :param src: source root directory